Update README.md
Committed to Change the Future for Better! Sadhana , Seva , Trust and Thoughts and Work .  

Thought for the Day: 

Earlier Days, People used to throw chocolate seeds and used to eat fruits only. A day came, and peope recognized that the Seeds are the source of tasty chacolates and later on Whole World, is enjoying chacolates and even a movie was made, on chocolates. Willy Wonka & The Chocolate Factory.